refactor(UserChannel) : UserChannel 엔티티, 레포지토리의 폴더 분리.
