+++ src/main/java/gdsc/comunity/entity/userchannel/UserChannel.java
@@ -1,15 +1,9 @@
-package gdsc.comunity.entity.user;
+package gdsc.comunity.entity.userchannel;
 
 import gdsc.comunity.entity.channel.Channel;
 import gdsc.comunity.entity.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import gdsc.comunity.entity.user.User;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
