+++ src/test/java/gdsc/comunity/service/channel/ChannelServiceImplTest.java
@@ -1,25 +1,20 @@
 package gdsc.comunity.service.channel;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.assertj.core.api.Assertions.assertThat;
-
 import gdsc.comunity.dto.channel.ChannelInfoDto;
 import gdsc.comunity.entity.channel.Channel;
 import gdsc.comunity.entity.channel.ChannelJoinRequest;
 import gdsc.comunity.entity.user.Provider;
 import gdsc.comunity.entity.user.User;
-import gdsc.comunity.entity.user.UserChannel;
+import gdsc.comunity.entity.userchannel.UserChannel;
 import gdsc.comunity.repository.channel.ChannelJoinRequestRepository;
 import gdsc.comunity.repository.channel.ChannelRepository;
-import gdsc.comunity.repository.user.UserChannelJpaRepository;
 import gdsc.comunity.repository.user.UserRepository;
+import gdsc.comunity.repository.userchannel.UserChannelJpaRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-//import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
@@ -27,6 +22,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest
